assignments/P01/requirements.txt
mongoengine==0.27.0
numpy==1.21.5
opencv_python==4.8.1.78
pandas==2.1.1
Pillow==9.0.1
Pillow==10.1.0
python_magic==0.4.27
rich==13.6.0
